+++ CHANGELOG.md
@@ -7,7 +7,9 @@ This change log adheres to standards from [Keep a CHANGELOG](http://keepachangel
 
 ### Fixed
 * [`no-invalid-html-attribute`]: allow `link` `rel` to have `apple-touch-icon`, `mask-icon` ([#3132][] @ljharb)
+* [`no-unused-class-component-methods`]: add `getChildContext` lifecycle method ([#3136][] @yoyo837)
 
+[#3136]: https://github.com/yannickcr/eslint-plugin-react/pull/3136
 [#3132]: https://github.com/yannickcr/eslint-plugin-react/issue/3132
 
 ## [7.27.0] - 2021.11.09